pyproject.toml
[project]
name = "fastapi-application"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "fastapi-users (>=14.0.1,<15.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "fastapi-pagination (>=0.14.1,<0.15.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "fastapi-users-db-sqlalchemy (>=7.0.0,<8.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "orjson (>=3.11.3,<4.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
